server/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-cors==4.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
requests==2.32.3
africastalking==1.2.8
python-dotenv==1.0.1
sqlalchemy==2.0.30